openstack-manifests: Refactor
