+++ scripts/openstack/manifest-tests/lb-default-stable/install-config.yaml
