+++ scripts/openstack/manifest-tests/lb-default-techpreview/install-config.yaml
