+++ scripts/openstack/manifest-tests/lb-managed/install-config.yaml
