+++ scripts/openstack/manifest-tests/lb-managed/test_cluster-infra.py
@@ -9,7 +9,7 @@ import yaml
 
 ASSETS_DIR = ""
 
-class TestClusterInfraObject(unittest.TestCase):
+class ManagedLoadBalancer(unittest.TestCase):
     def setUp(self):
         """Parse the Cluster Infrastructure object into a Python data structure."""
         self.machines = []
@@ -17,9 +17,14 @@ class TestClusterInfraObject(unittest.TestCase):
         with open(cluster_infra) as f:
             self.cluster_infra = yaml.load(f, Loader=yaml.FullLoader)
 
-    def test_load_balancer(self):
-        """Assert that the Cluster infrastructure object does not contain the LoadBalancer configuration."""
-        self.assertNotIn("loadBalancer", self.cluster_infra["status"]["platformStatus"]["openstack"])
+    def test_cluster_infra_object(self):
+        """Assert that the Cluster infrastructure object contains the LoadBalancer configuration."""
+        self.assertIn("loadBalancer", self.cluster_infra["status"]["platformStatus"]["openstack"])
+
+        loadBalancer = self.cluster_infra["status"]["platformStatus"]["openstack"]["loadBalancer"]
+
+        self.assertIn("type", loadBalancer)
+        self.assertEqual("OpenShiftManagedDefault", loadBalancer["type"])
 
 
 if __name__ == '__main__':
