+++ scripts/openstack/manifest-tests/server-groups/test_machines.py
@@ -10,7 +10,7 @@ import yaml
 ASSETS_DIR = ""
 
 
-class TestMachinesServerGroup(unittest.TestCase):
+class ServerGroupMachines(unittest.TestCase):
     def setUp(self):
         """Parse the Machines into a Python data structure."""
         self.machines = []
@@ -29,6 +29,26 @@ class TestMachinesServerGroup(unittest.TestCase):
                 group_name = name
 
             self.assertEqual(name, group_name)
+
+
+class ServerGroupMachinesets(unittest.TestCase):
+    def setUp(self):
+        """Parse the MachineSets into a Python data structure."""
+        self.machinesets = []
+        for machineset_path in glob.glob(
+                f'{ASSETS_DIR}/openshift/99_openshift-cluster-api_worker-machineset-*.yaml'
+        ):
+            with open(machineset_path) as f:
+                self.machinesets.append(yaml.load(f, Loader=yaml.FullLoader))
+
+    def test_consistent_group_names(self):
+        """Assert that server group names are unique across machinesets."""
+        found = []
+        for machineset in self.machinesets:
+            name = machineset["spec"]["template"]["spec"]["providerSpec"][
+                "value"]["serverGroupName"]
+            self.assertNotIn(name, found)
+            found.append(name)
 
 
 if __name__ == '__main__':
